feat: MainEntry 구현 완료 / update 수정 X
